chore(code): Move grouping code to own util
- Drop limit completely ref: 8c9a89033f771c0dd9da02052d22125be76fb3ba
- Add tid and wid to grouping logic
